// src/screens/MainTabScreens/Grading/interface.ts
export interface LessonItemProps {
    scheduleStartDate: string;
    sch_time: string | null;
    currDesc: string;
    lessonDesc: string;
    studentId: string;
    deviceId: string;
    traineeName: string;
    ste_date_complete: string | null;
    scheduleEndDate: string | null;
    rs_scale_id: number;
    rs_task_rating_id: number;
    ebt_scale_id: number;
    rs_reason_scale_id: number;
    level4b: number;
    lstIndicator: string;
    gateIndicator: string;
    le_mdetails_linetra: string;
    curriculumType: string;
    equipmentId: number;
    currId: number;
    lessonId: number;
    trainingCourseId: string;
    staffNo: string;
    supportCrew: string | null;
    otherInstrutor1: number;
    otherInstrutor2: number;
    trainingStatus: string;
    course_code: string | null;
    observerName: string | null;
    otherInstructorName1: string | null;
    otherInstructorName2: string | null;
    le_details_crewed: number;
    instructorId: string;
    aims_courseCode: string | null;
    trainee_password: string;
    lessonAbbr: string;
    sch_code: string;
    arival_airport: string | null;
    dep_airport: string | null;
    currAbbr: string;
    template_id: number;
  }